Consulta 7 realizada en el frontend y arrelgo de procedimiento almacenado actualizar_mantenimiento_por_modelo

+++ Backend/src/main/java/com/Lab1BDA/Backend/repository/actualizar_mantenimiento_por_modelo.sql
@@ -2,7 +2,7 @@
 -- Requisito 7 del enunciado: Actualización Masiva de Estado de Mantenimiento
 
 CREATE OR REPLACE FUNCTION actualizar_mantenimiento_por_modelo(
-    p_id_modelo INT
+    p_id_modelo BIGINT
 )
 RETURNS INTEGER AS $$ -- Devuelve el número de drones actualizados
 DECLARE
